--- project/src/main/java/com/google/sps/data/FakeSearchRequest.java
@@ -1,22 +1,0 @@
-package com.google.sps.data;
-
-import com.google.maps.TextSearchRequest;
-import com.google.maps.GeoApiContext;
-
-public class FakeSearchRequest extends TextSearchRequest {
-
-    /** The search wirds used for the text search */
-    public String searchWords;
-
-    /**
-     * FakeTestSearchRequest constructor.
-     *
-     * @param geoApiContext the GeoApiContext used for all Google GEO API requests
-     * @param searchWords the search words used for the text search
-     */
-    public FakeSearchRequest(GeoApiContext context, String searchWords) {
-        super(context);
-        this.searchWords = searchWords;
-    }
-
-}
