+++ project/src/main/java/com/google/sps/servlets/QueryServlet.java
@@ -32,10 +32,7 @@ import com.google.sps.data.PlacesFetcher;
 import com.google.sps.data.PlacesScorer;
 import com.google.sps.data.UserPreferences;
 import com.google.sps.data.PlacesScorerImpl;
-import com.google.sps.data.SearchRequestGenerator;
-import com.google.sps.data.SearchRequestGeneratorImpl;
-import com.google.sps.data.PlaceDetailsRequestGenerator;
-import com.google.sps.data.PlaceDetailsRequestGeneratorImpl;
+
 /**
  * A servlet that handles the user's food-mood recommendation query, and responds with a list of
  * recommended places (in Json format).
@@ -46,16 +43,12 @@ public final class QueryServlet extends HttpServlet {
 
   @VisibleForTesting
   static final int MAX_NUM_PLACES_TO_RECOMMEND = 3;
-  static final SearchRequestGenerator SEARCH_REQUEST_GENERATOR =
-      new SearchRequestGeneratorImpl(GeoContext.getGeoApiContext());
-  static final PlaceDetailsRequestGenerator PLACE_DETAILS_REQUEST_GENERATOR =
-      new PlaceDetailsRequestGeneratorImpl(GeoContext.getGeoApiContext());
   private PlacesFetcher fetcher;
   private PlacesScorer scorer;
 
   @Override
   public void init() {
-    fetcher = new PlacesFetcher(SEARCH_REQUEST_GENERATOR, PLACE_DETAILS_REQUEST_GENERATOR);
+    fetcher = new PlacesFetcher(GeoContext.getGeoApiContext());
     scorer = new PlacesScorerImpl(GeoContext.getGeoApiContext());
   }
