+++ project/src/test/java/com/google/sps/data/FakePlaceDetailsRequestGenerator.java
@@ -29,4 +29,22 @@ public class FakePlaceDetailsRequestGenerator implements PlaceDetailsRequestGene
         request.placeId(placeId);
         return request;
     }
+
+    public static class FakePlaceDetailsRequest extends PlaceDetailsRequest {
+
+        /** The Place ID to retrieve details for. */
+        public String placeId;
+
+        /**
+         * FakeTestSearchRequest constructor.
+         *
+         * @param geoApiContext the GeoApiContext used for all Google GEO API requests
+         * @param placeId the Place ID to retrieve details for
+         */
+        public FakePlaceDetailsRequest(GeoApiContext context, String placeId) {
+            super(context);
+            this.placeId = placeId;
+        }
+    }
+
 }
